solutions/quickback(Flask+Postgres)/src/python/requirements.txt
flask
flask-cors
cryptography
pyjwt
requests
gunicorn
psycopg2-binary